feat: SupplierInfoをDouble型に変更、ClassCategoryにisActive追加
- FoodRawMaterialSupplier: purchasePrice/volumeAmountをDouble型に変更
- ClassCategory: isActiveフィールドを追加しDB保存/取得対応
- ApixFoodRawMaterialSupplierController: Double型対応メソッド追加
- ApixMasterClassCategoryController: isActive対応
- ClassCategoryMapper.xml: is_activeカラム追加

+++ src/main/java/com/nines/nutsfact/domain/model/FoodRawMaterialSupplier.java
@@ -19,11 +19,11 @@ public class FoodRawMaterialSupplier {
     private Integer foodId;
     private Integer supplierId;
     private String supplierName;
-    private Integer purchasePrice;
-    private Integer volumeAmount;
+    private Double purchasePrice;
+    private Double volumeAmount;
     private Integer weightOrCapacity;
-    private Float convertRatio;
-    private Float pricePerUnit;
+    private Double convertRatio;
+    private Double pricePerUnit;
     private Date createDate;
     private Date lastUpdateDate;
     private Boolean isActive;
